requirements.txt
Django==1.6.2
MySQL-python==1.2.3
South==1.0.2
argparse==1.3.0
boto==2.28.0
httplib2==0.9
mercurial==2.9.1
path.py==5.1
requests==2.2.1
stevedore==0.15
virtualenv==1.11.4
virtualenv-clone==0.2.4
virtualenvwrapper==4.2
wsgiref==0.1.2
django-crispy-forms==1.4.0